Sign-Up (#1)
## Purpose of changes
Adding a sing up page validates the username, email and password. After it's been validated, the new user is saved into the database and the nececarry documents (under users, friends and locations) are created.

## Notes
The application is currently trying to redirect to "/map" route after a successful register which fails, as this route has not been created yet.

Co-authored-by: gukkame <guka245@gmail.com>
Co-authored-by: Laura-Eliise <laura-eliis@marrandi.ee>
Reviewed-on: https://01.kood.tech/git/Laura-Eliise/kahoot/pulls/1