Login (#2)
## Purpose of changes
Adding a sing in page which validates the email and password. After it's been validated it redirect the user into the map view, otherwise it informs the user of the error.

## Notes
The application is currently trying to redirect to "/map" route after a successful register which fails, as this route has not been created yet.

Co-authored-by: gukkame <guka245@gmail.com>
Co-authored-by: Laura-Eliise <laura-eliis@marrandi.ee>
Reviewed-on: https://01.kood.tech/git/Laura-Eliise/kahoot/pulls/2